Renames Blog to About and adds AboutGrid component
This revision refactors the Blog component and renames it to the About component while also implementing an about.css file instead of blog.css. It introduces the AboutGrid component into the GRIDS constant array and incorporates it into the HomeGrid component. The LandingPage component now also uses about.css instead of blog.css.